Make TextChannel implementation-independent
