+++ src/main/java/de/btobastian/javacord/entity/channel/TextChannel.java
@@ -1,20 +1,12 @@
 package de.btobastian.javacord.entity.channel;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.fasterxml.jackson.databind.node.JsonNodeFactory;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import de.btobastian.javacord.DiscordApi;
-import de.btobastian.javacord.ImplDiscordApi;
 import de.btobastian.javacord.entity.message.Message;
 import de.btobastian.javacord.entity.message.MessageSet;
 import de.btobastian.javacord.entity.message.Messageable;
-import de.btobastian.javacord.entity.message.impl.ImplMessageSet;
 import de.btobastian.javacord.entity.permission.PermissionType;
 import de.btobastian.javacord.entity.user.User;
 import de.btobastian.javacord.entity.webhook.Webhook;
-import de.btobastian.javacord.entity.webhook.impl.ImplWebhook;
-import de.btobastian.javacord.listener.ChannelAttachableListener;
 import de.btobastian.javacord.listener.ObjectAttachableListener;
 import de.btobastian.javacord.listener.TextChannelAttachableListener;
 import de.btobastian.javacord.listener.message.CachedMessagePinListener;
@@ -27,20 +19,14 @@ import de.btobastian.javacord.listener.message.reaction.ReactionAddListener;
 import de.btobastian.javacord.listener.message.reaction.ReactionRemoveAllListener;
 import de.btobastian.javacord.listener.message.reaction.ReactionRemoveListener;
 import de.btobastian.javacord.listener.user.UserStartTypingListener;
-import de.btobastian.javacord.util.ClassHelper;
 import de.btobastian.javacord.util.NonThrowingAutoCloseable;
 import de.btobastian.javacord.util.cache.MessageCache;
 import de.btobastian.javacord.util.event.ListenerManager;
 import de.btobastian.javacord.util.logging.LoggerUtil;
-import de.btobastian.javacord.util.rest.RestEndpoint;
-import de.btobastian.javacord.util.rest.RestMethod;
-import de.btobastian.javacord.util.rest.RestRequest;
 import org.slf4j.Logger;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
@@ -50,8 +36,6 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
-import java.util.stream.LongStream;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
@@ -76,12 +60,7 @@ public interface TextChannel extends Channel, Messageable {
      * @see #typeContinuously()
      * @see #typeContinuously(Consumer)
      */
-    default CompletableFuture<Void> type() {
-        return new RestRequest<Void>(getApi(), RestMethod.POST, RestEndpoint.CHANNEL_TYPING)
-                .setRatelimitRetries(0)
-                .setUrlParameters(getIdAsString())
-                .execute(result -> null);
-    }
+    CompletableFuture<Void> type();
 
     /**
      * Displays the "xyz is typing..." message continuously, starting immediately.
@@ -222,7 +201,7 @@ public interface TextChannel extends Channel, Messageable {
                 });
 
         // auto-closable to cancel the continuously typing indicator
-        return  () -> {
+        return () -> {
             typingInterruptedListenerManager.remove();
             typingIndicator.cancel(true);
         };
@@ -250,19 +229,7 @@ public interface TextChannel extends Channel, Messageable {
      * @param messageIds The ids of the messages to delete.
      * @return A future to tell us if the deletion was successful.
      */
-    default CompletableFuture<Void> bulkDelete(long... messageIds) {
-        ObjectNode body = JsonNodeFactory.instance.objectNode();
-        ArrayNode messages = body.putArray("messages");
-        LongStream.of(messageIds).boxed()
-                .map(Long::toUnsignedString)
-                .forEach(messages::add);
-
-        return new RestRequest<Void>(getApi(), RestMethod.POST, RestEndpoint.MESSAGES_BULK_DELETE)
-                .setRatelimitRetries(0)
-                .setUrlParameters(getIdAsString())
-                .setBody(body)
-                .execute(result -> null);
-    }
+    CompletableFuture<Void> bulkDelete(long... messageIds);
 
     /**
      * Deletes multiple messages at once.
@@ -366,13 +333,7 @@ public interface TextChannel extends Channel, Messageable {
      * @param id The id of the message.
      * @return The message with the given id.
      */
-    default CompletableFuture<Message> getMessageById(long id) {
-        return getApi().getCachedMessageById(id)
-                .map(CompletableFuture::completedFuture)
-                .orElseGet(() -> new RestRequest<Message>(getApi(), RestMethod.GET, RestEndpoint.MESSAGE)
-                .setUrlParameters(getIdAsString(), Long.toUnsignedString(id))
-                .execute(result -> ((ImplDiscordApi) getApi()).getOrCreateMessage(this, result.getJsonBody())));
-    }
+    CompletableFuture<Message> getMessageById(long id);
 
     /**
      * Gets a message by its id.
@@ -393,16 +354,7 @@ public interface TextChannel extends Channel, Messageable {
      *
      * @return A list with all pinned messages.
      */
-    default CompletableFuture<MessageSet> getPins() {
-        return new RestRequest<MessageSet>(getApi(), RestMethod.GET, RestEndpoint.PINS)
-                .setUrlParameters(getIdAsString())
-                .execute(result -> {
-                    Collection<Message> pins = StreamSupport.stream(result.getJsonBody().spliterator(), false)
-                            .map(pinJson -> ((ImplDiscordApi) getApi()).getOrCreateMessage(this, pinJson))
-                            .collect(Collectors.toList());
-                    return new ImplMessageSet(pins);
-                });
-    }
+    CompletableFuture<MessageSet> getPins();
 
     /**
      * Gets up to a given amount of messages in this channel from the newer end.
@@ -411,9 +363,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAsStream()
      */
-    default CompletableFuture<MessageSet> getMessages(int limit) {
-        return ImplMessageSet.getMessages(this, limit);
-    }
+    CompletableFuture<MessageSet> getMessages(int limit);
 
     /**
      * Gets messages in this channel from the newer end until one that meets the given condition is found.
@@ -423,9 +373,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAsStream()
      */
-    default CompletableFuture<MessageSet> getMessagesUntil(Predicate<Message> condition) {
-        return ImplMessageSet.getMessagesUntil(this, condition);
-    }
+    CompletableFuture<MessageSet> getMessagesUntil(Predicate<Message> condition);
 
     /**
      * Gets messages in this channel from the newer end while they meet the given condition.
@@ -435,9 +383,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAsStream()
      */
-    default CompletableFuture<MessageSet> getMessagesWhile(Predicate<Message> condition) {
-        return ImplMessageSet.getMessagesWhile(this, condition);
-    }
+    CompletableFuture<MessageSet> getMessagesWhile(Predicate<Message> condition);
 
     /**
      * Gets a stream of messages in this channel sorted from newest to oldest.
@@ -448,9 +394,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The stream.
      * @see #getMessages(int)
      */
-    default Stream<Message> getMessagesAsStream() {
-        return ImplMessageSet.getMessagesAsStream(this);
-    }
+    Stream<Message> getMessagesAsStream();
 
     /**
      * Gets up to a given amount of messages in this channel before a given message in any channel.
@@ -460,9 +404,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesBeforeAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesBefore(int limit, long before) {
-        return ImplMessageSet.getMessagesBefore(this, limit, before);
-    }
+    CompletableFuture<MessageSet> getMessagesBefore(int limit, long before);
 
     /**
      * Gets messages in this channel before a given message in any channel until one that meets the given condition is
@@ -474,9 +416,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesBeforeAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesBeforeUntil(Predicate<Message> condition, long before) {
-        return ImplMessageSet.getMessagesBeforeUntil(this, condition, before);
-    }
+    CompletableFuture<MessageSet> getMessagesBeforeUntil(Predicate<Message> condition, long before);
 
     /**
      * Gets messages in this channel before a given message in any channel while they meet the given condition.
@@ -487,9 +427,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesBeforeAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesBeforeWhile(Predicate<Message> condition, long before) {
-        return ImplMessageSet.getMessagesBeforeWhile(this, condition, before);
-    }
+    CompletableFuture<MessageSet> getMessagesBeforeWhile(Predicate<Message> condition, long before);
 
     /**
      * Gets a stream of messages in this channel before a given message in any channel sorted from newest to oldest.
@@ -501,9 +439,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The stream.
      * @see #getMessagesBefore(int, long)
      */
-    default Stream<Message> getMessagesBeforeAsStream(long before) {
-        return ImplMessageSet.getMessagesBeforeAsStream(this, before);
-    }
+    Stream<Message> getMessagesBeforeAsStream(long before);
 
     /**
      * Gets up to a given amount of messages in this channel before a given message in any channel.
@@ -566,9 +502,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAfterAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesAfter(int limit, long after) {
-        return ImplMessageSet.getMessagesAfter(this, limit, after);
-    }
+    CompletableFuture<MessageSet> getMessagesAfter(int limit, long after);
 
     /**
      * Gets messages in this channel after a given message in any channel until one that meets the given condition is
@@ -580,9 +514,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAfterAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesAfterUntil(Predicate<Message> condition, long after) {
-        return ImplMessageSet.getMessagesAfterUntil(this, condition, after);
-    }
+    CompletableFuture<MessageSet> getMessagesAfterUntil(Predicate<Message> condition, long after);
 
     /**
      * Gets messages in this channel after a given message in any channel while they meet the given condition.
@@ -593,9 +525,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAfterAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesAfterWhile(Predicate<Message> condition, long after) {
-        return ImplMessageSet.getMessagesAfterWhile(this, condition, after);
-    }
+    CompletableFuture<MessageSet> getMessagesAfterWhile(Predicate<Message> condition, long after);
 
     /**
      * Gets a stream of messages in this channel after a given message in any channel sorted from oldest to newest.
@@ -607,9 +537,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAfter(int, long)
      */
-    default Stream<Message> getMessagesAfterAsStream(long after) {
-        return ImplMessageSet.getMessagesAfterAsStream(this, after);
-    }
+    Stream<Message> getMessagesAfterAsStream(long after);
 
     /**
      * Gets up to a given amount of messages in this channel after a given message in any channel.
@@ -677,9 +605,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAroundAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesAround(int limit, long around) {
-        return ImplMessageSet.getMessagesAround(this, limit, around);
-    }
+    CompletableFuture<MessageSet> getMessagesAround(int limit, long around);
 
     /**
      * Gets messages in this channel around a given message in any channel until one that meets the given condition is
@@ -695,9 +621,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAroundAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesAroundUntil(Predicate<Message> condition, long around) {
-        return ImplMessageSet.getMessagesAroundUntil(this, condition, around);
-    }
+    CompletableFuture<MessageSet> getMessagesAroundUntil(Predicate<Message> condition, long around);
 
     /**
      * Gets messages in this channel around a given message in any channel while they meet the given condition.
@@ -713,9 +637,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesAroundAsStream(long)
      */
-    default CompletableFuture<MessageSet> getMessagesAroundWhile(Predicate<Message> condition, long around) {
-        return ImplMessageSet.getMessagesAroundWhile(this, condition, around);
-    }
+    CompletableFuture<MessageSet> getMessagesAroundWhile(Predicate<Message> condition, long around);
 
     /**
      * Gets a stream of messages in this channel around a given message in any channel.
@@ -731,9 +653,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The stream.
      * @see #getMessagesAround(int, long)
      */
-    default Stream<Message> getMessagesAroundAsStream(long around) {
-        return ImplMessageSet.getMessagesAroundAsStream(this, around);
-    }
+    Stream<Message> getMessagesAroundAsStream(long around);
 
     /**
      * Gets up to a given amount of messages in this channel around a given message in any channel.
@@ -815,9 +735,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesBetweenAsStream(long, long)
      */
-    default CompletableFuture<MessageSet> getMessagesBetween(long from, long to) {
-        return ImplMessageSet.getMessagesBetween(this, from, to);
-    }
+    CompletableFuture<MessageSet> getMessagesBetween(long from, long to);
 
     /**
      * Gets all messages in this channel between the first given message in any channel and the second given message in
@@ -830,9 +748,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesBetweenAsStream(long, long)
      */
-    default CompletableFuture<MessageSet> getMessagesBetweenUntil(Predicate<Message> condition, long from, long to) {
-        return ImplMessageSet.getMessagesBetweenUntil(this, condition, from, to);
-    }
+    CompletableFuture<MessageSet> getMessagesBetweenUntil(Predicate<Message> condition, long from, long to);
 
     /**
      * Gets all messages in this channel between the first given message in any channel and the second given message in
@@ -845,9 +761,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The messages.
      * @see #getMessagesBetweenAsStream(long, long)
      */
-    default CompletableFuture<MessageSet> getMessagesBetweenWhile(Predicate<Message> condition, long from, long to) {
-        return ImplMessageSet.getMessagesBetweenWhile(this, condition, from, to);
-    }
+    CompletableFuture<MessageSet> getMessagesBetweenWhile(Predicate<Message> condition, long from, long to);
 
     /**
      * Gets all messages in this channel between the first given message in any channel and the second given message in
@@ -861,9 +775,7 @@ public interface TextChannel extends Channel, Messageable {
      * @return The stream.
      * @see #getMessagesBetween(long, long)
      */
-    default Stream<Message> getMessagesBetweenAsStream(long from, long to) {
-        return ImplMessageSet.getMessagesBetweenAsStream(this, from, to);
-    }
+    Stream<Message> getMessagesBetweenAsStream(long from, long to);
 
     /**
      * Gets all messages in this channel between the first given message in any channel and the second given message in
@@ -938,17 +850,7 @@ public interface TextChannel extends Channel, Messageable {
      *
      * @return A list of all webhooks in this channel.
      */
-    default CompletableFuture<List<Webhook>> getWebhooks() {
-        return new RestRequest<List<Webhook>>(getApi(), RestMethod.GET, RestEndpoint.CHANNEL_WEBHOOK)
-                .setUrlParameters(getIdAsString())
-                .execute(result -> {
-                    List<Webhook> webhooks = new ArrayList<>();
-                    for (JsonNode webhook : result.getJsonBody()) {
-                        webhooks.add(new ImplWebhook(getApi(), webhook));
-                    }
-                    return Collections.unmodifiableList(webhooks);
-                });
-    }
+    CompletableFuture<List<Webhook>> getWebhooks();
 
     /**
      * Checks if the given user can send messages in this channel.
@@ -970,9 +872,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
-                || severTextChannel.get()
-                .hasPermissions(user, PermissionType.READ_MESSAGES, PermissionType.SEND_MESSAGES);
+               || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
+               || severTextChannel.get()
+                       .hasPermissions(user, PermissionType.READ_MESSAGES, PermissionType.SEND_MESSAGES);
     }
 
     /**
@@ -1003,9 +905,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasAnyPermission(user,
-                                                           PermissionType.ADMINISTRATOR,
-                                                           PermissionType.USE_EXTERNAL_EMOJIS);
+               || severTextChannel.get().hasAnyPermission(user,
+                                                          PermissionType.ADMINISTRATOR,
+                                                          PermissionType.USE_EXTERNAL_EMOJIS);
     }
 
     /**
@@ -1036,9 +938,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasAnyPermission(user,
-                                                           PermissionType.ADMINISTRATOR,
-                                                           PermissionType.EMBED_LINKS);
+               || severTextChannel.get().hasAnyPermission(user,
+                                                          PermissionType.ADMINISTRATOR,
+                                                          PermissionType.EMBED_LINKS);
     }
 
     /**
@@ -1067,9 +969,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasAnyPermission(user,
-                                                           PermissionType.ADMINISTRATOR,
-                                                           PermissionType.READ_MESSAGE_HISTORY);
+               || severTextChannel.get().hasAnyPermission(user,
+                                                          PermissionType.ADMINISTRATOR,
+                                                          PermissionType.READ_MESSAGE_HISTORY);
     }
 
     /**
@@ -1098,9 +1000,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasAnyPermission(user,
-                                                           PermissionType.ADMINISTRATOR,
-                                                           PermissionType.SEND_TTS_MESSAGES);
+               || severTextChannel.get().hasAnyPermission(user,
+                                                          PermissionType.ADMINISTRATOR,
+                                                          PermissionType.SEND_TTS_MESSAGES);
     }
 
     /**
@@ -1132,9 +1034,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
-                || (severTextChannel.get().hasPermission(user, PermissionType.ATTACH_FILE)
-                && severTextChannel.get().canWrite(user));
+               || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
+               || (severTextChannel.get().hasPermission(user, PermissionType.ATTACH_FILE)
+                   && severTextChannel.get().canWrite(user));
     }
 
     /**
@@ -1163,11 +1065,11 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
-                || severTextChannel.get().hasPermissions(user,
-                PermissionType.READ_MESSAGES,
-                PermissionType.READ_MESSAGE_HISTORY,
-                PermissionType.ADD_REACTIONS);
+               || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
+               || severTextChannel.get().hasPermissions(user,
+                                                        PermissionType.READ_MESSAGES,
+                                                        PermissionType.READ_MESSAGE_HISTORY,
+                                                        PermissionType.ADD_REACTIONS);
     }
 
     /**
@@ -1193,9 +1095,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasAnyPermission(user,
-                                                           PermissionType.ADMINISTRATOR,
-                                                           PermissionType.MANAGE_MESSAGES);
+               || severTextChannel.get().hasAnyPermission(user,
+                                                          PermissionType.ADMINISTRATOR,
+                                                          PermissionType.MANAGE_MESSAGES);
     }
 
     /**
@@ -1247,9 +1149,9 @@ public interface TextChannel extends Channel, Messageable {
         }
         Optional<ServerTextChannel> severTextChannel = asServerTextChannel();
         return !severTextChannel.isPresent()
-                || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
-                || (severTextChannel.get().hasPermission(user, PermissionType.MENTION_EVERYONE)
-                && severTextChannel.get().canWrite(user));
+               || severTextChannel.get().hasPermission(user, PermissionType.ADMINISTRATOR)
+               || (severTextChannel.get().hasPermission(user, PermissionType.MENTION_EVERYONE)
+                   && severTextChannel.get().canWrite(user));
     }
 
     /**
@@ -1269,19 +1171,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<MessageCreateListener> addMessageCreateListener(MessageCreateListener listener) {
-        return ((ImplDiscordApi) getApi()).addObjectListener(
-                TextChannel.class, getId(), MessageCreateListener.class, listener);
-    }
+    ListenerManager<MessageCreateListener> addMessageCreateListener(MessageCreateListener listener);
 
     /**
      * Gets a list with all registered message create listeners.
      *
      * @return A list with all registered message create listeners.
      */
-    default List<MessageCreateListener> getMessageCreateListeners() {
-        return ((ImplDiscordApi) getApi()).getObjectListeners(TextChannel.class, getId(), MessageCreateListener.class);
-    }
+    List<MessageCreateListener> getMessageCreateListeners();
 
     /**
      * Adds a listener, which listens to users starting to type in this channel.
@@ -1289,20 +1186,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<UserStartTypingListener> addUserStartTypingListener(UserStartTypingListener listener) {
-        return ((ImplDiscordApi) getApi()).addObjectListener(
-                TextChannel.class, getId(), UserStartTypingListener.class, listener);
-    }
+    ListenerManager<UserStartTypingListener> addUserStartTypingListener(UserStartTypingListener listener);
 
     /**
      * Gets a list with all registered user starts typing listeners.
      *
      * @return A list with all registered user starts typing listeners.
      */
-    default List<UserStartTypingListener> getUserStartTypingListeners() {
-        return ((ImplDiscordApi) getApi()).getObjectListeners(
-                TextChannel.class, getId(), UserStartTypingListener.class);
-    }
+    List<UserStartTypingListener> getUserStartTypingListeners();
 
     /**
      * Adds a listener, which listens to message deletions in this channel.
@@ -1310,19 +1201,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<MessageDeleteListener> addMessageDeleteListener(MessageDeleteListener listener) {
-        return ((ImplDiscordApi) getApi()).addObjectListener(
-                TextChannel.class, getId(), MessageDeleteListener.class, listener);
-    }
+    ListenerManager<MessageDeleteListener> addMessageDeleteListener(MessageDeleteListener listener);
 
     /**
      * Gets a list with all registered message delete listeners.
      *
      * @return A list with all registered message delete listeners.
      */
-    default List<MessageDeleteListener> getMessageDeleteListeners() {
-        return ((ImplDiscordApi) getApi()).getObjectListeners(TextChannel.class, getId(), MessageDeleteListener.class);
-    }
+    List<MessageDeleteListener> getMessageDeleteListeners();
 
     /**
      * Adds a listener, which listens to message edits in this channel.
@@ -1330,19 +1216,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<MessageEditListener> addMessageEditListener(MessageEditListener listener) {
-        return ((ImplDiscordApi) getApi())
-                .addObjectListener(TextChannel.class, getId(), MessageEditListener.class, listener);
-    }
+    ListenerManager<MessageEditListener> addMessageEditListener(MessageEditListener listener);
 
     /**
      * Gets a list with all registered message edit listeners.
      *
      * @return A list with all registered message edit listeners.
      */
-    default List<MessageEditListener> getMessageEditListeners() {
-        return ((ImplDiscordApi) getApi()).getObjectListeners(TextChannel.class, getId(), MessageEditListener.class);
-    }
+    List<MessageEditListener> getMessageEditListeners();
 
     /**
      * Adds a listener, which listens to reactions being added in this channel.
@@ -1350,19 +1231,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<ReactionAddListener> addReactionAddListener(ReactionAddListener listener) {
-        return ((ImplDiscordApi) getApi())
-                .addObjectListener(TextChannel.class, getId(), ReactionAddListener.class, listener);
-    }
+    ListenerManager<ReactionAddListener> addReactionAddListener(ReactionAddListener listener);
 
     /**
      * Gets a list with all registered reaction add listeners.
      *
      * @return A list with all registered reaction add listeners.
      */
-    default List<ReactionAddListener> getReactionAddListeners() {
-        return ((ImplDiscordApi) getApi()).getObjectListeners(TextChannel.class, getId(), ReactionAddListener.class);
-    }
+    List<ReactionAddListener> getReactionAddListeners();
 
     /**
      * Adds a listener, which listens to reactions being removed in this channel.
@@ -1370,19 +1246,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<ReactionRemoveListener> addReactionRemoveListener(ReactionRemoveListener listener) {
-        return ((ImplDiscordApi) getApi()).addObjectListener(
-                TextChannel.class, getId(), ReactionRemoveListener.class, listener);
-    }
+    ListenerManager<ReactionRemoveListener> addReactionRemoveListener(ReactionRemoveListener listener);
 
     /**
      * Gets a list with all registered reaction remove listeners.
      *
      * @return A list with all registered reaction remove listeners.
      */
-    default List<ReactionRemoveListener> getReactionRemoveListeners() {
-        return ((ImplDiscordApi) getApi()).getObjectListeners(TextChannel.class, getId(), ReactionRemoveListener.class);
-    }
+    List<ReactionRemoveListener> getReactionRemoveListeners();
 
     /**
      * Adds a listener, which listens to all reactions being removed at once from a message in this channel.
@@ -1390,21 +1261,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<ReactionRemoveAllListener> addReactionRemoveAllListener(
-            ReactionRemoveAllListener listener) {
-        return ((ImplDiscordApi) getApi())
-                .addObjectListener(TextChannel.class, getId(), ReactionRemoveAllListener.class, listener);
-    }
+    ListenerManager<ReactionRemoveAllListener> addReactionRemoveAllListener(ReactionRemoveAllListener listener);
 
     /**
      * Gets a list with all registered reaction remove all listeners.
      *
      * @return A list with all registered reaction remove all listeners.
      */
-    default List<ReactionRemoveAllListener> getReactionRemoveAllListeners() {
-        return ((ImplDiscordApi) getApi())
-                .getObjectListeners(TextChannel.class, getId(), ReactionRemoveAllListener.class);
-    }
+    List<ReactionRemoveAllListener> getReactionRemoveAllListeners();
 
     /**
      * Adds a listener, which listens to all pin updates in this channel.
@@ -1412,21 +1276,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<ChannelPinsUpdateListener> addChannelPinsUpdateListener(
-            ChannelPinsUpdateListener listener) {
-        return ((ImplDiscordApi) getApi())
-                .addObjectListener(TextChannel.class, getId(), ChannelPinsUpdateListener.class, listener);
-    }
+    ListenerManager<ChannelPinsUpdateListener> addChannelPinsUpdateListener(ChannelPinsUpdateListener listener);
 
     /**
      * Gets a list with all registered channel pins update listeners.
      *
      * @return A list with all registered channel pins update listeners.
      */
-    default List<ChannelPinsUpdateListener> getChannelPinsUpdateListeners() {
-        return ((ImplDiscordApi) getApi())
-                .getObjectListeners(TextChannel.class, getId(), ChannelPinsUpdateListener.class);
-    }
+    List<ChannelPinsUpdateListener> getChannelPinsUpdateListeners();
 
     /**
      * Adds a listener, which listens to all cached message pins in this channel.
@@ -1434,20 +1291,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<CachedMessagePinListener> addCachedMessagePinListener(CachedMessagePinListener listener) {
-        return ((ImplDiscordApi) getApi())
-                .addObjectListener(TextChannel.class, getId(), CachedMessagePinListener.class, listener);
-    }
+    ListenerManager<CachedMessagePinListener> addCachedMessagePinListener(CachedMessagePinListener listener);
 
     /**
      * Gets a list with all registered cached message pin listeners.
      *
      * @return A list with all registered cached message pin listeners.
      */
-    default List<CachedMessagePinListener> getCachedMessagePinListeners() {
-        return ((ImplDiscordApi) getApi())
-                .getObjectListeners(TextChannel.class, getId(), CachedMessagePinListener.class);
-    }
+    List<CachedMessagePinListener> getCachedMessagePinListeners();
 
     /**
      * Adds a listener, which listens to all cached message unpins in this channel.
@@ -1455,21 +1306,14 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to add.
      * @return The manager of the listener.
      */
-    default ListenerManager<CachedMessageUnpinListener> addCachedMessageUnpinListener(
-            CachedMessageUnpinListener listener) {
-        return ((ImplDiscordApi) getApi())
-                .addObjectListener(TextChannel.class, getId(), CachedMessageUnpinListener.class, listener);
-    }
+    ListenerManager<CachedMessageUnpinListener> addCachedMessageUnpinListener(CachedMessageUnpinListener listener);
 
     /**
      * Gets a list with all registered cached message unpin listeners.
      *
      * @return A list with all registered cached message unpin listeners.
      */
-    default List<CachedMessageUnpinListener> getCachedMessageUnpinListeners() {
-        return ((ImplDiscordApi) getApi())
-                .getObjectListeners(TextChannel.class, getId(), CachedMessageUnpinListener.class);
-    }
+    List<CachedMessageUnpinListener> getCachedMessageUnpinListeners();
 
     /**
      * Adds a listener that implements one or more {@code TextChannelAttachableListener}s.
@@ -1482,24 +1326,8 @@ public interface TextChannel extends Channel, Messageable {
      * @return The managers for the added listener.
      */
     @SuppressWarnings("unchecked")
-    default <T extends TextChannelAttachableListener & ObjectAttachableListener>
-    Collection<ListenerManager<? extends TextChannelAttachableListener>>
-    addTextChannelAttachableListener(T listener) {
-        return ClassHelper.getInterfacesAsStream(listener.getClass())
-                .filter(TextChannelAttachableListener.class::isAssignableFrom)
-                .filter(ObjectAttachableListener.class::isAssignableFrom)
-                .map(listenerClass -> (Class<T>) listenerClass)
-                .flatMap(listenerClass -> {
-                    if (ChannelAttachableListener.class.isAssignableFrom(listenerClass)) {
-                        return addChannelAttachableListener(
-                                (ChannelAttachableListener & ObjectAttachableListener) listener).stream();
-                    } else {
-                        return Stream.of(((ImplDiscordApi) getApi()).addObjectListener(TextChannel.class, getId(),
-                                                                                       listenerClass, listener));
-                    }
-                })
-                .collect(Collectors.toList());
-    }
+    <T extends TextChannelAttachableListener & ObjectAttachableListener>
+    Collection<ListenerManager<? extends TextChannelAttachableListener>> addTextChannelAttachableListener(T listener);
 
     /**
      * Removes a listener that implements one or more {@code TextChannelAttachableListener}s.
@@ -1508,22 +1336,8 @@ public interface TextChannel extends Channel, Messageable {
      * @param <T> The type of the listener.
      */
     @SuppressWarnings("unchecked")
-    default <T extends TextChannelAttachableListener & ObjectAttachableListener> void
-    removeTextChannelAttachableListener(T listener) {
-        ClassHelper.getInterfacesAsStream(listener.getClass())
-                .filter(TextChannelAttachableListener.class::isAssignableFrom)
-                .filter(ObjectAttachableListener.class::isAssignableFrom)
-                .map(listenerClass -> (Class<T>) listenerClass)
-                .forEach(listenerClass -> {
-                    if (ChannelAttachableListener.class.isAssignableFrom(listenerClass)) {
-                        removeChannelAttachableListener(
-                                (ChannelAttachableListener & ObjectAttachableListener) listener);
-                    } else {
-                        ((ImplDiscordApi) getApi()).removeObjectListener(TextChannel.class, getId(),
-                                                                         listenerClass, listener);
-                    }
-                });
-    }
+    <T extends TextChannelAttachableListener & ObjectAttachableListener> void removeTextChannelAttachableListener(
+            T listener);
 
     /**
      * Gets a map with all registered listeners that implement one or more {@code TextChannelAttachableListener}s and
@@ -1534,19 +1348,8 @@ public interface TextChannel extends Channel, Messageable {
      * their assigned listener classes they listen to.
      */
     @SuppressWarnings("unchecked")
-    default <T extends TextChannelAttachableListener & ObjectAttachableListener> Map<T, List<Class<T>>>
-    getTextChannelAttachableListeners() {
-        Map<T, List<Class<T>>> textChannelListeners =
-                ((ImplDiscordApi) getApi()).getObjectListeners(TextChannel.class, getId());
-        getChannelAttachableListeners().forEach((listener, listenerClasses) -> textChannelListeners
-                .merge((T) listener,
-                       (List<Class<T>>) (Object) listenerClasses,
-                       (listenerClasses1, listenerClasses2) -> {
-                           listenerClasses1.addAll(listenerClasses2);
-                           return listenerClasses1;
-                       }));
-        return textChannelListeners;
-    }
+    <T extends TextChannelAttachableListener & ObjectAttachableListener> Map<T, List<Class<T>>>
+    getTextChannelAttachableListeners();
 
     /**
      * Removes a listener from this text channel.
@@ -1555,10 +1358,8 @@ public interface TextChannel extends Channel, Messageable {
      * @param listener The listener to remove.
      * @param <T> The type of the listener.
      */
-    default <T extends TextChannelAttachableListener & ObjectAttachableListener> void removeListener(
-            Class<T> listenerClass, T listener) {
-        ((ImplDiscordApi) getApi()).removeObjectListener(TextChannel.class, getId(), listenerClass, listener);
-    }
+    <T extends TextChannelAttachableListener & ObjectAttachableListener> void removeListener(
+            Class<T> listenerClass, T listener);
 
     @Override
     default Optional<? extends TextChannel> getCurrentCachedInstance() {
