+++ src/main/java/de/btobastian/javacord/entity/channel/impl/ImplGroupChannel.java
@@ -8,6 +8,7 @@ import de.btobastian.javacord.entity.Icon;
 import de.btobastian.javacord.entity.channel.GroupChannel;
 import de.btobastian.javacord.entity.channel.GroupChannelUpdater;
 import de.btobastian.javacord.entity.channel.InternalChannel;
+import de.btobastian.javacord.entity.channel.InternalTextChannel;
 import de.btobastian.javacord.entity.impl.ImplIcon;
 import de.btobastian.javacord.entity.user.User;
 import de.btobastian.javacord.listener.ChannelAttachableListener;
@@ -40,7 +41,7 @@ import java.util.stream.Stream;
 /**
  * The implementation of {@link GroupChannel}.
  */
-public class ImplGroupChannel implements GroupChannel, Cleanupable, InternalChannel {
+public class ImplGroupChannel implements GroupChannel, Cleanupable, InternalChannel, InternalTextChannel {
 
     /**
      * The logger of this class.
