+++ src/main/java/de/btobastian/javacord/entity/channel/impl/ImplPrivateChannel.java
@@ -5,6 +5,7 @@ import de.btobastian.javacord.DiscordApi;
 import de.btobastian.javacord.ImplDiscordApi;
 import de.btobastian.javacord.entity.DiscordEntity;
 import de.btobastian.javacord.entity.channel.InternalChannel;
+import de.btobastian.javacord.entity.channel.InternalTextChannel;
 import de.btobastian.javacord.entity.channel.PrivateChannel;
 import de.btobastian.javacord.entity.user.User;
 import de.btobastian.javacord.entity.user.impl.ImplUser;
@@ -30,7 +31,7 @@ import java.util.stream.Stream;
 /**
  * The implementation of {@link PrivateChannel}.
  */
-public class ImplPrivateChannel implements PrivateChannel, Cleanupable, InternalChannel {
+public class ImplPrivateChannel implements PrivateChannel, Cleanupable, InternalChannel, InternalTextChannel {
 
     /**
      * The discord api instance.
