+++ src/main/java/de/btobastian/javacord/entity/channel/impl/ImplServerTextChannel.java
@@ -7,6 +7,7 @@ import de.btobastian.javacord.entity.DiscordEntity;
 import de.btobastian.javacord.entity.channel.ChannelCategory;
 import de.btobastian.javacord.entity.channel.InternalChannel;
 import de.btobastian.javacord.entity.channel.InternalServerChannel;
+import de.btobastian.javacord.entity.channel.InternalTextChannel;
 import de.btobastian.javacord.entity.channel.ServerTextChannel;
 import de.btobastian.javacord.entity.channel.ServerTextChannelUpdater;
 import de.btobastian.javacord.entity.permission.Permissions;
@@ -43,7 +44,8 @@ import java.util.stream.Stream;
 /**
  * The implementation of {@link ServerTextChannel}.
  */
-public class ImplServerTextChannel implements ServerTextChannel, Cleanupable, InternalChannel, InternalServerChannel {
+public class ImplServerTextChannel
+        implements ServerTextChannel, Cleanupable, InternalChannel, InternalServerChannel, InternalTextChannel {
 
     /**
      * The discord api instance.
